+++ src/library/CustomHooks/SignUpCustomHook.js
@@ -4,12 +4,13 @@ export const SignUpCustomHook = () => {
     
     const [disable, setDisable] = useState(true);
     const [name, setName] = useState(false);
-    const [email, setEmail] = useState(false);
+    const [email, setEcmail] = useState(false);
     const [emailText, setEmailText] = useState('');
     const [password, setPassword] = useState(false);
     const [passwordText, setPasswordText] = useState('');
     const [terms, setTerms] = useState(false);
     const [subscribe, setSubscribe] = useState(false);
+    
     const handleText = text => {
         if(text.length >= 1){
             setEmail(true);
@@ -28,25 +29,6 @@ export const SignUpCustomHook = () => {
             setDisable(false);
         }
     }
-    return {
-        disable,
-        setDisable,
-        name,
-        setName,
-        email,
-        setEmail,
-        emailText,
-        setEmailText,
-        password,
-        setPassword,
-        passwordText,
-        setPasswordText,
-        terms,
-        setTerms,
-        subscribe,
-        setSubscribe,
-        handleText,
-        handlePassword,
-        handleCheckbox
-    }
+    return { disable, setDisable, name, setName, email, setEmail, emailText, setEmailText, password, setPassword, passwordText, setPasswordText,
+        terms, setTerms, subscribe, setSubscribe, handleText, handlePassword, handleCheckbox }
 }
